Ajustes nas implementacoes

+++ src/main/java/com/marcospedroso/facens/correlato/dto/create/CreateUpdateCurso.java
@@ -19,5 +19,5 @@ public class CreateUpdateCurso {
     private String nome;
 	
 	@NotNull(message = "não pode ser nulo")
-	private int idFaculdade;
+	private Long idFaculdade;
 }
